requirements/common.txt
django==1.8.4

pytz
Jinja2==2.7.3
django-jinja==1.4.1
dj-database-url==0.2.2
# TODO (yml): Remove hstore when we squash the migration
django-hstore==1.2.5
# django-statsd-mozilla==0.3.12
git+https://github.com/vbabiy/django-statsd.git@request-aggregation

markdown==2.3.1
redis==2.9.1
django-pipeline==1.5.1
    # Requires java runtime
    yuicompressor==2.4.8
honcho==0.5.0

djorm-ext-pgfulltext==0.9.0
djorm-ext-pgarray==0.9.0

django-allauth==0.20.0
    requests==2.7.0
    oauthlib==0.7.2
    requests-oauthlib==0.5.0
django-bootstrap-toolkit==2.15.0

geoip2==2.1.0

# Needs pinning
-e git+https://github.com/lincolnloop/django-jsonit.git#egg=jsonit
-e git+https://github.com/lincolnloop/django-launchpad.git#egg=launchpad

# For plugins
-e git+https://github.com/BotBotMe/botbot-plugins.git#egg=botbot-plugins

